Index: source/internal/rules/design/PublicImplementationRule.cs * Only fire if none of the new public methods are called.
Index: source/internal/rules/design/UnusedMethodRule.cs
* Fixed (it only worked half the time).

Index: source/tests/rules/design/EqualityOperatorTest.cs
Index: source/internal/rules/design/EqualityOperatorRule.cs
* Ignore private structs.

Index: source/internal/CheckItem.cs
* CompilerGenerated methods are now checked so that anonymous delegates are
  properly processed. This means that rules may need to special case these types.
  
Index: source/tests/rules/performance/UseEmptyTypesTest.cs
Index: source/internal/rules/performance/UseEmptyTypesRule.cs
* Fixed so that it only fires if the type is "System.Type".

Index: install.sh
* Doesn't overwrite smokey.exe.config if it already exists.

Index: source/app/AnalyzeAssembly.cs
Index: source/framework/support/MethodMatcher.cs
Index: source/framework/StringExtensions.cs
Index: source/framework/support/CecilExtensions.cs
Index: source/framework/support/advanced/ClassifyMethod.cs
Index: source/framework/support/AssemblyCache.cs
* Added some asserts.

Index: source/internal/Program.cs
* Main no longer returns negative ints.

Index: source/tests/V0.cs
...
* Sealed lots and lots of classes.

+++ trunk/install.sh
@@ -10,15 +10,17 @@ cp bin/smokey.exe "$BIN_DIR"
 chmod -x "$BIN_DIR/smokey.exe"
 
 # Write a hard-coded config file alongside the assembly.
-echo "<?xml version = \"1.0\" encoding = \"utf-8\" ?>" > "$BIN_DIR/smokey.exe.config"
-echo "<configuration>" >> "$BIN_DIR/smokey.exe.config"
-echo "	<appSettings>" >> "$BIN_DIR/smokey.exe.config"
-echo "		<add key = \"logfile\" value = \"$LOG_FILE\"/>" >> "$BIN_DIR/smokey.exe.config"
-echo "		<add key = \"topic:System.Object\" value = \"Info\"/>	<!-- may be off, Error, Warning, Info, Trace, or Debug -->" >> "$BIN_DIR/smokey.exe.config"
-echo "		<add key = \"consoleWidth\" value = \"80\"/>			<!-- TextReport breaks lines so that that they aren't longer than this -->" >> "$BIN_DIR/smokey.exe.config"
-echo "		<add key = \"topic:System.Object\" value = \"Warning\"/>	" >> "$BIN_DIR/smokey.exe.config"
-echo "	</appSettings>" >> "$BIN_DIR/smokey.exe.config"
-echo "</configuration>" >> "$BIN_DIR/smokey.exe.config"
+if [ ! -f "$BIN_DIR/smokey.exe.config" ]; then
+	echo "<?xml version = \"1.0\" encoding = \"utf-8\" ?>" > "$BIN_DIR/smokey.exe.config"
+	echo "<configuration>" >> "$BIN_DIR/smokey.exe.config"
+	echo "	<appSettings>" >> "$BIN_DIR/smokey.exe.config"
+	echo "		<add key = \"logfile\" value = \"$LOG_FILE\"/>" >> "$BIN_DIR/smokey.exe.config"
+	echo "		<add key = \"topic:System.Object\" value = \"Info\"/>	<!-- may be off, Error, Warning, Info, Trace, or Debug -->" >> "$BIN_DIR/smokey.exe.config"
+	echo "		<add key = \"consoleWidth\" value = \"80\"/>			<!-- TextReport breaks lines so that that they aren't longer than this -->" >> "$BIN_DIR/smokey.exe.config"
+	echo "		<add key = \"topic:System.Object\" value = \"Warning\"/>	" >> "$BIN_DIR/smokey.exe.config"
+	echo "	</appSettings>" >> "$BIN_DIR/smokey.exe.config"
+	echo "</configuration>" >> "$BIN_DIR/smokey.exe.config"
+fi
 
 # Create a little smokey launcher script.
 echo "#!/bin/sh" > "$BIN_DIR/smoke"
